+++ CHANGELOG.md
@@ -1,4 +1,22 @@
 ## FlatPlayer+ ChangeLog
 
-### [1.0.0]
+## [2.0.0] Major Update
+
+### Added
+* BoneMenu
+
+* HUD
+
+* Preferences
+
+* Grip Toggle
+
+* Some Graphics Settings
+    * FOV (Field of view)
+    * Full-Screen Mode
+    * V-Sync
+
+* Customizable Bindings
+
+## [1.0.0]  Regular Update
 * Initial Release